se modificaro todas las ramas del proyecto

+++ src/vehiculosinterfases/Motor.java
@@ -3,7 +3,7 @@ package vehiculosinterfases;
 
 
 public interface Motor {
-    int calcularRevolucionesMotor(int fuerza, int radio);
+    double calcularRevolucionesMotor(int fuerza, int radio);
     
     
 }
